ATTOX

+++ include/stdint.h
@@ -5,7 +5,7 @@ typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
 typedef unsigned long long uint64_t;
-typedef unsigned long uintptr_t;
+typedef uint32_t uintptr_t;
 typedef unsigned int size_t;
 
 #endif
